src/requirements.txt
alembic==1.8.1
amqp==5.1.1
Babel==2.10.3
bcrypt==3.2.2
bleach==5.0.1
blinker==1.5
cachelib==0.9.0
celery==5.2.7
cffi==1.15.1
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Caching==2.0.0
Flask-DebugToolbar==0.13.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security-Too==4.1.5
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.1
MarkupSafe==2.1.1
passlib==1.7.4
Pillow==9.2.0
psycopg2-binary==2.9.3
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.20.0
python-editor==1.0.4
pytz==2022.1
redis==4.3.4
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.39
#uWSGI==2.0.20
vine==5.0.0
Werkzeug==2.2.1
WTForms==3.0.1
BS4==0.0.1
nltk==3.7
scipy==1.9.1
lxml==4.9.1
pandas==1.5.0
gensim==4.2.0
interrogate==1.5.0